requirements.txt
fastapi>=0.100.0
uvicorn>=0.20.0
chromadb>=0.4.0
requests>=2.31.0
pyyaml>=6.0
pywin32>=306
selenium>=4.15.0
webdriver-manager>=4.0.1
